fix(libsinsp): Reset destination IP on old connect enter event
When a connect enter event comes from a driver that does not
report the connection tuple argument, reset the destination
address to zero. We know the old destination is no longer valid
but don't know the new one until the exit event arrives
so a buch of zeros is the best we can have.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>